Fix type mismatch in transaction-line count fields
Cast transaction_line_count and null_date_count to UInt32 when creating
pl.lit() expressions. The schema expects UInt32 but pl.lit() infers Int32
by default, causing SchemaError on extend().

Fixes: 'type Int32 is incompatible with expected type UInt32' error

+++ src/bank_statement_parser/modules/statements.py
@@ -499,8 +499,8 @@ class Statement:
                     .otherwise(pl.lit(0))
                 ).sum().item() or 0
                 self.checks_and_balances = self.checks_and_balances.with_columns(
-                    TRANSACTION_LINE_COUNT=pl.lit(transaction_line_count),
-                    TRANSACTION_LINES_WITH_NULL_DATE=pl.lit(null_date_count),
+                    TRANSACTION_LINE_COUNT=pl.lit(transaction_line_count, dtype=pl.UInt32),
+                    TRANSACTION_LINES_WITH_NULL_DATE=pl.lit(null_date_count, dtype=pl.UInt32),
                 )
             self.logs.rechunk()
             self.success = self.is_successfull()
